Restructure credentials binding conroller directories
