+++ skaffold-operator.yaml
@@ -663,6 +663,7 @@ build:
             - pkg/controllermanager/controller/controllerregistration/seed
             - pkg/controllermanager/controller/controllerregistration/seedfinalizer
             - pkg/controllermanager/controller/credentialsbinding
+            - pkg/controllermanager/controller/credentialsbinding/credentialsbinding
             - pkg/controllermanager/controller/credentialsbinding/referencecleaner
             - pkg/controllermanager/controller/event
             - pkg/controllermanager/controller/exposureclass
